incident_management/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.2
django-extensions==3.2.3
django-mysql==4.15.0
djangorestframework==3.15.2
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
